Refactor FakeNumericDataset into separate file: dataset_fakes.ts
